refactor(mentor): 파일 정리
